change oriented

+++ 0x08-recursion/main.h
@@ -7,4 +7,4 @@ int _strlen_recursion(char *s);
 int factorial(int n);
 int _pow_recursion(int x, int y);
 int _sqrt_recursion(int n);
-int squrt(int, int n, int);
+int squrt(long, long n, long);
